Implemented argument options.  Changed Max role logic
Argument options are now parsed and set.
Only autoStart is not implemented, since the feature itself is still WiP.
Max Role logic changed to behave akin to old bot, anything below 0 is unlimited, 0 disables the role, above is a limit.  It is recommended that unneeded roles are not added and instead removed.