Commander Update
Connections updated to use event triggers instead of an interval; now changes to voice states or, if a ps2 event, a player login/logout causes an update.

User library also updated to use new Session data layout.

Session data objects updated to include __version, which can later be used when reading a session if its outdated and can no longer be displayed (or delete it)